backend/requirements-streamlit.txt
streamlit==1.32.0
chromadb==0.4.22
pandas==2.2.0
python-dateutil==2.8.2
uuid==1.30
Pillow==10.2.0
plotly==5.18.0
pydantic==2.6.0
